# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Seattle Solvers Tools API",
    "description": "FTC parts inventory, search, and receipt reading for tools.seattlesolvers.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/parts": {
      "get": {
        "summary": "Search the parts library (global + your team's custom parts)",
        "tags": [
          "parts"
        ],
        "description": "Anonymous callers see the approved global library. Authenticated callers also see their team's custom parts and an `ownedQuantity` for each result.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "category",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "manufacturer",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "GLOBAL",
                "TEAM"
              ]
            },
            "in": "query",
            "name": "scope",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "ownedOnly",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 100,
              "default": 25
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "sku": {
                            "type": "string",
                            "nullable": true
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "productUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "purchaseUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "imageUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "scope": {
                            "type": "string",
                            "enum": [
                              "GLOBAL",
                              "TEAM"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "APPROVED",
                              "PENDING",
                              "REJECTED"
                            ]
                          },
                          "manufacturer": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "slug"
                            ],
                            "additionalProperties": false
                          },
                          "category": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "slug"
                            ],
                            "additionalProperties": false
                          },
                          "createdByTeamId": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "ownedQuantity": {
                            "type": "integer",
                            "nullable": true
                          },
                          "lastKnownPrice": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "sku",
                          "description",
                          "productUrl",
                          "purchaseUrl",
                          "imageUrl",
                          "scope",
                          "status",
                          "manufacturer",
                          "category",
                          "createdByTeamId",
                          "createdAt",
                          "lastKnownPrice"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize",
                    "total",
                    "totalPages"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a custom part for your team, optionally requesting it for the shared library",
        "tags": [
          "parts"
        ],
        "description": "Always creates a part scoped to your team, usable immediately — a personal inventory entry never needs more than a name, manufacturer, and category, so nothing here blocks on that. `submitToLibrary` additionally *requests* the part for the shared global library instead of adding it there outright: it requires a product URL (every approved library part must stay traceable), and is skipped rather than queued when `duplicateOf` already reports a matching or pending part — this is the guard against duplicate and junk submissions piling up in Seattle Solvers' review queue. A real request still needs SUPER_ADMIN approval (GET/POST /admin/submissions) before it reaches every team.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "sku": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "productUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "purchaseUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "manufacturerId": {
                    "type": "string"
                  },
                  "categoryId": {
                    "type": "string"
                  },
                  "unitCost": {
                    "type": "number",
                    "minimum": 0
                  },
                  "submitToLibrary": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "name",
                  "manufacturerId",
                  "categoryId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "part": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "sku": {
                          "type": "string",
                          "nullable": true
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "productUrl": {
                          "type": "string",
                          "nullable": true
                        },
                        "purchaseUrl": {
                          "type": "string",
                          "nullable": true
                        },
                        "imageUrl": {
                          "type": "string",
                          "nullable": true
                        },
                        "scope": {
                          "type": "string",
                          "enum": [
                            "GLOBAL",
                            "TEAM"
                          ]
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "APPROVED",
                            "PENDING",
                            "REJECTED"
                          ]
                        },
                        "manufacturer": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "slug"
                          ],
                          "additionalProperties": false
                        },
                        "category": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "slug"
                          ],
                          "additionalProperties": false
                        },
                        "createdByTeamId": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "ownedQuantity": {
                          "type": "integer",
                          "nullable": true
                        },
                        "lastKnownPrice": {
                          "type": "number",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "sku",
                        "description",
                        "productUrl",
                        "purchaseUrl",
                        "imageUrl",
                        "scope",
                        "status",
                        "manufacturer",
                        "category",
                        "createdByTeamId",
                        "createdAt",
                        "lastKnownPrice"
                      ],
                      "additionalProperties": false
                    },
                    "submittedToLibrary": {
                      "type": "boolean"
                    },
                    "duplicateOf": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "APPROVED",
                            "PENDING",
                            "REJECTED"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "status"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "part",
                    "submittedToLibrary",
                    "duplicateOf"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/parts/suggest-url": {
      "get": {
        "summary": "Best-effort product page URL and image for a part not yet in the library",
        "tags": [
          "parts"
        ],
        "description": "Deterministic for vendors with a known SKU-to-URL formula (currently REV); otherwise a web-search-grounded guess from Claude, billed to your team's own Anthropic key (see PATCH /teams/current), or null if your team has not set one or Claude was not confident. Once a product URL is known, its page's own og:image is read automatically — reviewers never hand-paste an image URL. Always a suggestion to review, never a verified link. Feeds into adding a part, so it is gated the same way: TEAM_ADMIN only.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "GOBILDA",
                "REV",
                "AXON",
                "FERRA",
                "MELONBOTICS",
                "OFFSET",
                "MATA",
                "UXCELL",
                "OTHER"
              ]
            },
            "in": "query",
            "name": "vendor",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sku",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "name",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "deterministic",
                        "ai_search",
                        "none"
                      ]
                    },
                    "imageUrl": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true
                    }
                  },
                  "required": [
                    "url",
                    "source",
                    "imageUrl"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}